refactor(nvim): rename lspconfig.lua -> mason.lua
