+++ nvim/lua/plugins/configs/mason.lua
